+++ core/connection/tcp/src/softbus_tcp_connect_manager.c
@@ -293,7 +293,7 @@ int32_t TcpOnDataEventIn(int32_t fd)
         DelTcpConnInfo(connectionId);
         return SOFTBUS_ERR;
     }
-    g_tcpConnCallback->OnDataReceived(connectionId, head.module, head.seq, data, headSize + head.len);
+    g_tcpConnCallback->OnDataReceived(connectionId, head.module, head.seq, data, (int32_t)headSize + head.len);
     SoftBusFree(data);
     return SOFTBUS_OK;
 }
